tambah fitur autentikasi login

+++ app/Config/Routes.php
@@ -17,7 +17,7 @@ if (file_exists(SYSTEMPATH . 'Config/Routes.php')) {
  * --------------------------------------------------------------------
  */
 $routes->setDefaultNamespace('App\Controllers');
-$routes->setDefaultController('Home');
+$routes->setDefaultController('Login');
 $routes->setDefaultMethod('index');
 $routes->setTranslateURIDashes(false);
 $routes->set404Override();
@@ -31,22 +31,25 @@ $routes->setAutoRoute(true);
 
 // We get a performance increase by specifying the default
 // route since we don't have to scan directories.
-$routes->get('/', 'Beranda::index');
+$routes->get('/', 'Login::index');
+
 $routes->get('/login', 'Login::index');
 
-$routes->get('/menu', 'Menu::index');
-$routes->get('/menu/create', 'Menu::create');
-$routes->get('/menu/edit/(:segment)', 'Menu::edit/$1');
-$routes->delete('/menu/(:any)', 'Menu::delete/$1');
-$routes->get('/menu/(:any)','Menu::detail/$1');
+$routes->get('/beranda', 'Beranda::index', ['filter' => 'auth']);
 
-$routes->get('/pemesanan', 'Pemesanan::index');
-$routes->get('/laporan', 'Laporan::index');
-$routes->get('/pembayaran', 'Pembayaran::index');
-$routes->get('/hitung_bayar', 'Hitung_bayar::index');
-$routes->get('/pencarian_meja', 'Pemesanan::pencarian_meja');
-$routes->get('/tambah_pesanan', 'Pemesanan::tambah_pesanan');
-$routes->get('/ubah_pesanan', 'Pemesanan::ubah_pesanan');
+$routes->get('/menu', 'Menu::index', ['filter' => 'auth']);
+$routes->get('/menu/create', 'Menu::create', ['filter' => 'auth']);
+$routes->get('/menu/edit/(:segment)', 'Menu::edit/$1', ['filter' => 'auth']);
+$routes->delete('/menu/(:any)', 'Menu::delete/$1', ['filter' => 'auth']);
+$routes->get('/menu/(:any)','Menu::detail/$1', ['filter' => 'auth']);
+
+$routes->get('/pemesanan', 'Pemesanan::index', ['filter' => 'auth']);
+$routes->get('/laporan', 'Laporan::index', ['filter' => 'auth']);
+$routes->get('/pembayaran', 'Pembayaran::index', ['filter' => 'auth']);
+$routes->get('/hitung_bayar', 'Hitung_bayar::index', ['filter' => 'auth']);
+$routes->get('/pencarian_meja', 'Pemesanan::pencarian_meja', ['filter' => 'auth']);
+$routes->get('/tambah_pesanan', 'Pemesanan::tambah_pesanan', ['filter' => 'auth']);
+$routes->get('/ubah_pesanan', 'Pemesanan::ubah_pesanan', ['filter' => 'auth']);
 
 /*
  * --------------------------------------------------------------------
